% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplasso.R
\name{bplasso}
\alias{bplasso}
\title{Bio-primed LASSO}
\usage{
bplasso(X, y, scores, n_folds = 10, phi_range = seq(0, 1, length = 30))
}
\arguments{
\item{X}{matrix as in `glmnet::glmnet()`}

\item{y}{response array as in `glmnet::glmnet()`}

\item{scores}{Gene specific association scores, from `get_scores()`}

\item{n_folds}{Number of cross-validations}

\item{phi_range}{Array of phi ranges to test.  Suggested limits [0,1]}
}
\value{
list of 5 elements
1: "phi" = best phi value as chosen by `find_best_phi_rmse()`
2: "lambda" = lambda value as identified as `lambda.min` by `glmnet::cv.glmnet()`
3: "betas" = data.frame containing baseline lasso betas "betas" and bio-primed lasso betas "betas_pen" for all columns of matrix `x`
}
\description{
Bio-primed LASSO
}
\examples{
 bplasso(scale(X),
  y,
  scores,
  n_folds = 10,
  phi_range = seq(0, 1, length = 30))

}
