% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_phi_rmse.R
\name{find_best_phi_rmse}
\alias{find_best_phi_rmse}
\title{Identify phi value
Cross-validation across range of phi values, find best phi by RMSE}
\usage{
find_best_phi_rmse(rmse, phi_range)
}
\arguments{
\item{rmse}{Correlation matrix along phi value range from `get_rmse_from_xvalid`}

\item{phi_range}{Array of phi ranges to test.  Suggested limits [0,1]}
}
\value{
A numeric value
}
\description{
Identify phi value
Cross-validation across range of phi values, find best phi by RMSE
}
\examples{
phirange <- seq(0,1,length = 30)
rmse <- get_rmse_from_xvalid(X, y, penalties,
  phi_range = phirange,
  lambda_min = 0.5,
  n_folds = 10)
  best_phi <- find_best_phi_rmse(rmse, phirange, plot = F)
}
