% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lambda.R
\name{find_lambda}
\alias{find_lambda}
\title{Calculate best lambda: lambda min}
\usage{
find_lambda(X, y, plot = F)
}
\arguments{
\item{X}{matrix as in `glmnet`}

\item{y}{response array as in `glmnet::glmnet()`}

\item{plot}{plot the resulting cross-validation for glmnet}
}
\value{
A numeric value
}
\description{
Calculate best lambda: lambda min
}
