% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmse_from_xvalid.R
\name{get_rmse_from_xvalid}
\alias{get_rmse_from_xvalid}
\title{RSME for the range of lambda
RMSE, via N-fold cross-validation, along a range of phi, given lambda.  Returned as a Z-score}
\usage{
get_rmse_from_xvalid(X, y, penalties, lambda_min, phi_range, n_folds = 10)
}
\arguments{
\item{X}{matrix as in `glmnet::glmnet()`}

\item{y}{response array as in `glmnet::glmnet()`}

\item{penalties}{User-supplied penalty scores.  Range [0-1]}

\item{lambda_min}{User-supplied lambda value from `find_lambda()`}

\item{phi_range}{Array of phi ranges to test.  Suggested limits [0,1]}

\item{n_folds}{Number of cross-validations}
}
\value{
a matrix of n-fold by phi_range z-scored RMSE, from cross-validation glmnet
}
\description{
RSME for the range of lambda
RMSE, via N-fold cross-validation, along a range of phi, given lambda.  Returned as a Z-score
}
\examples{
#' phirange <- seq(0,1,length = 30)
  get_rmse_from_xvalid(X, y, penalties,
    phi_range = phirange,
    lambda_min = 0.5,
    n_folds = 10)

}
