% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores}
\alias{get_scores}
\title{Gene specific association scores}
\usage{
get_scores(gene, network)
}
\arguments{
\item{gene}{Gene symbol. e.g. "EGFR"}

\item{network}{Data frame of parameter-pair scores (range: [0-1])}
}
\value{
A numerical vector named
}
\description{
Gene specific association scores
}
\examples{

head(network)
         protein1        protein2 combined_score gene1    gene2
1 ENSP00000000233 ENSP00000356607            0.173  ARF5  RALGPS2
2 ENSP00000000233 ENSP00000427567            0.154  ARF5    FHDC1
3 ENSP00000000233 ENSP00000253413            0.151  ARF5 ATP6V1E1
...

scores <- get_scores(gene="EGFR", network)
head(scores)
ENPEP  LOXL3   SYT1 FCGR2B   AKT1   RARB
0.265  0.197  0.228  0.275  0.792  0.420
}
