% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plots of bplasso results}
\usage{
plot_manhattan(gene, resIn, subplotChr = NA, dependency, gene_info, dir_save)
}
\arguments{
\item{gene}{Gene symbol. e.g. "EGFR"}

\item{resIn}{Path to "results_omic".RData file.}

\item{subplotChr}{Chromosome number for subplot}

\item{dependency}{Dependency data frame which includes gene column}

\item{gene_info}{Gene information data frame. Columns should include : hgnc_symbol, chromosome_name, start_position}

\item{dir_save}{directory in which plots will be saved}
}
\value{
Three Manhattan plots are written into the dir_save folder
}
\description{
Manhattan plots of bplasso results
}
\examples{
Gene <- "EGFR"
results_EGFR <- "../Outputs/EGFR_demeter2_CNV.RData"
eleven <- 11
D2 <- demeter2
dir_save <- "../Outputs/Graphics/"
plot_manhattan(gene=Gene,resIn=results_EGFR,subplotChr=eleven,dependency=D2,dir_save=fldr_plts)
}
